Move utility classes to utils folder
