+++ app/static/js/utils/Utils.js
